+++ engine/docs/changelog.md
@@ -1,4 +1,9 @@
 # TBD
+### Features
+* Upgraded to Kurt Core v1.34.0, which allows users to add user-friendly port IDs to the ports that they declare for their containers
+
+### Removals
+* Removed the `ListenProtocol` engine server arg, because it listens on gRPC and only TCP is supported
 
 # 1.5.6
 ### Fixes